rename NodeDefinition to NodeDefinitions
