+++ packages/core/src/Nodes/FlowNode.ts
@@ -3,7 +3,7 @@ import { Fiber } from '../Execution/Fiber';
 import { IGraphApi } from '../Graphs/Graph';
 import { Socket } from '../Sockets/Socket';
 import { Node, NodeConfiguration } from './Node';
-import { NodeCategory } from './NodeDefinition';
+import { NodeCategory } from './NodeDefinitions';
 import { IFlowNode, NodeType } from './NodeInstance';
 import { NodeDescription } from './Registry/NodeDescription';
