+++ packages/core/src/Profiles/Core/Flow/Branch.ts
@@ -1,7 +1,7 @@
 import {
   makeFlowNodeDefinition,
   NodeCategory
-} from '../../../Nodes/NodeDefinition';
+} from '../../../Nodes/NodeDefinitions';
 
 export const Branch = makeFlowNodeDefinition({
   typeName: 'flow/branch',
