+++ packages/core/src/Profiles/Core/Flow/Counter.ts
@@ -1,7 +1,7 @@
 import {
   makeFlowNodeDefinition,
   NodeCategory
-} from '../../../Nodes/NodeDefinition';
+} from '../../../Nodes/NodeDefinitions';
 
 export const Counter = makeFlowNodeDefinition({
   typeName: 'flow/counter',
