+++ packages/core/src/Profiles/Core/Lifecycle/LifecycleOnEnd.ts
@@ -2,7 +2,7 @@ import { Assert } from '../../../Diagnostics/Assert';
 import {
   makeEventNodeDefinition,
   NodeCategory
-} from '../../../Nodes/NodeDefinition';
+} from '../../../Nodes/NodeDefinitions';
 import { ILifecycleEventEmitter } from '../Abstractions/ILifecycleEventEmitter';
 
 // inspired by: https://docs.unrealengine.com/4.27/en-US/ProgrammingAndScripting/Blueprints/UserGuide/Events/
