+++ packages/core/src/Profiles/Core/Variables/VariableSet.ts
@@ -2,7 +2,7 @@ import {
   makeFlowNodeDefinition,
   NodeCategory,
   SocketsList
-} from '../../../Nodes/NodeDefinition';
+} from '../../../Nodes/NodeDefinitions';
 import { Variable } from '../../../Variables/Variable';
 
 export const VariableSet = makeFlowNodeDefinition({
